pyproject.toml
[tool.poetry]
name = "timelessnesses-api-qrbarpy"
version = "0.1.0"
description = "A QR code/barcode HTTP API written in Python and uses FastAPI and uvicorn"
authors = ["timelessnesses <timelessnesses@timelessnesses.me>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.95.2"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
qrcode = {extras = ["pil"], version = "^7.4.2"}
python-barcode = {extras = ["images"], version = "^0.14.0"}
python-multipart = "^0.0.6"
opencv-python = "^4.7.0.72"
pyzbar = "^0.1.9"
orjson = "^3.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
